# drop karrick/godirwalk in favour of stdlib