@mzihlmann 2025-08-13
let's check both performance and size for this. It looks like performance is not really relevant on that codepath, a fairly large build snapshot ~ 1min only spends a few seconds on discovery:

before
```
WARN[0166] elapsed: 2.708731663s                        
WARN[0168] elapsed: 2.434525402s                        
WARN[0171] elapsed: 2.270565059s                        
WARN[0173] elapsed: 2.448614492s                        
WARN[0175] elapsed: 2.164136475s                        
WARN[0177] elapsed: 2.182288801s 
```
2.368 +- 0.206

after
```
WARN[0167] elapsed: 2.289002816s                        
WARN[0169] elapsed: 2.077576024s                        
WARN[0172] elapsed: 2.315977054s                        
WARN[0174] elapsed: 2.259141514s                        
WARN[0176] elapsed: 2.052964085s                        
WARN[0178] elapsed: 2.333455137s 
```
2.221 +- 0.124

so performance wise we would even see a slight improvement. This is inline with what others reported about swapping out this library having reached EOL. https://engineering.kablamo.com.au/posts/quick-comparison-between-go-file-walk-implementations/index.html

Size wise let's see how much we save:
before
```
-rwxr-xr-x    1 0        0          11.1M Aug 13 18:05 docker-credential-acr-env
-rwxr-xr-x    1 0        0          12.3M Aug 13 18:05 docker-credential-ecr-login
-rwxr-xr-x    1 0        0          10.6M Aug 13 18:05 docker-credential-gcr
-rwxr-xr-x    1 0        0          48.5M Aug 13 18:44 executor
drwxr-xr-x    3 0        0           4.0K Aug 11 20:44 ssl
-rwxr-xr-x    1 0        0          12.9M Aug 13 18:44 warmer
```

after
```
-rwxr-xr-x    1 0        0          11.1M Aug 13 18:05 docker-credential-acr-env
-rwxr-xr-x    1 0        0          12.3M Aug 13 18:05 docker-credential-ecr-login
-rwxr-xr-x    1 0        0          10.6M Aug 13 18:05 docker-credential-gcr
-rwxr-xr-x    1 0        0          48.5M Aug 13 18:06 executor
drwxr-xr-x    3 0        0           4.0K Aug 11 20:44 ssl
-rwxr-xr-x    1 0        0          12.9M Aug 13 18:06 warmer
```
unfortunately no change there, the library is just too small.